+++ configs/omarchy/extensions/menu.sh
@@ -3,20 +3,18 @@
 #
 # WARNING: Overwritten functions will obviously not be updated when Omarchy changes.
 #
-# Force shutdown - kills all apps immediately without waiting
-show_system_menu() {
-  local options="  Lock\n󱄄  Screensaver"
-  [ -f ~/.local/state/omarchy/toggles/suspend-on ] && options="$options\n󰒲  Suspend"
-  omarchy-hibernation-available && options="$options\n󰤁  Hibernate"
-  options="$options\n󰜉  Restart\n󰐥  Shutdown"
-
-  case $(menu "System" "$options") in
-  *Lock*) omarchy-lock-screen ;;
-  *Screensaver*) omarchy-launch-screensaver force ;;
-  *Suspend*) systemctl suspend ;;
-  *Hibernate*) systemctl hibernate ;;
-  *Restart*) shutdown -r now ;;
-  *Shutdown*) shutdown now ;;
-  *) back_to show_main_menu ;;
-  esac
-}
+# Example of minimal system menu:
+#
+# show_system_menu() {
+#   case $(menu "System" "  Lock\n󰐥  Shutdown") in
+#   *Lock*) omarchy-system-lock ;;
+#   *Shutdown*) omarchy-system-shutdown ;;
+#   *) back_to show_main_menu ;;
+#   esac
+# }
+#
+# Example of overriding just the about menu action: (Using zsh instead of bash (default))
+#
+# show_about() {
+#   exec omarchy-launch-or-focus-tui "zsh -c 'fastfetch; read -k 1'"
+# }